+++ src/test/java/com/commercetools/sync/categories/helpers/CategoryReferenceResolverTest.java
@@ -29,7 +29,9 @@ import java.util.concurrent.CompletableFuture;
 
 import static com.commercetools.sync.categories.CategorySyncMockUtils.getMockCategoryDraftBuilder;
 import static com.commercetools.sync.commons.MockUtils.getMockTypeService;
+import static com.commercetools.sync.commons.helpers.BaseReferenceResolver.BLANK_ID_VALUE_ON_RESOURCE_IDENTIFIER;
 import static io.sphere.sdk.models.LocalizedString.ofEnglish;
+import static java.lang.String.format;
 import static java.util.Collections.emptyList;
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.singletonList;
@@ -191,7 +193,6 @@ public class CategoryReferenceResolverTest {
                                  .thenAccept(resolvedDraft -> {
                                      assertThat(resolvedDraft.getParent()).isNotNull();
                                      assertThat(resolvedDraft.getParent().getId()).isEqualTo(CACHED_CATEGORY_ID);
-                                     assertThat(resolvedDraft.getParent().getObj()).isNull();
                                  }).toCompletableFuture().join();
     }
 
@@ -250,32 +251,30 @@ public class CategoryReferenceResolverTest {
     public void resolveParentReference_WithEmptyIdOnParentReference_ShouldNotResolveParentReference() {
         final CategoryDraftBuilder categoryDraft = CategoryDraftBuilder.of(ofEnglish("foo"), ofEnglish("bar"));
         categoryDraft.key("key");
-        categoryDraft.parent(Category.referenceOfId(""));
+        categoryDraft.parent(Category.referenceOfId("").toResourceIdentifier());
 
         assertThat(referenceResolver.resolveParentReference(categoryDraft)
             .toCompletableFuture())
             .hasFailed()
             .hasFailedWithThrowableThat()
             .isExactlyInstanceOf(ReferenceResolutionException.class)
-            .hasMessage("Failed to resolve parent reference on CategoryDraft"
-                + " with key:'key'. Reason: Key is blank (null/empty) on both expanded"
-                + " reference object and reference id field.");
+            .hasMessage(format("Failed to resolve parent reference on CategoryDraft with key:'key'. Reason: %s",
+                BLANK_ID_VALUE_ON_RESOURCE_IDENTIFIER));
     }
 
     @Test
     public void resolveParentReference_WithNullIdOnParentReference_ShouldNotResolveParentReference() {
         final CategoryDraftBuilder categoryDraft = CategoryDraftBuilder.of(ofEnglish("foo"), ofEnglish("bar"));
         categoryDraft.key("key");
-        categoryDraft.parent(Category.referenceOfId(null));
+        categoryDraft.parent(Category.referenceOfId(null).toResourceIdentifier());
 
         assertThat(referenceResolver.resolveParentReference(categoryDraft)
             .toCompletableFuture())
             .hasFailed()
             .hasFailedWithThrowableThat()
             .isExactlyInstanceOf(ReferenceResolutionException.class)
-            .hasMessage("Failed to resolve parent reference on CategoryDraft"
-                + " with key:'key'. Reason: Key is blank (null/empty) on both expanded"
-                + " reference object and reference id field.");
+            .hasMessage(format("Failed to resolve parent reference on CategoryDraft with key:'key'. Reason: %s",
+                BLANK_ID_VALUE_ON_RESOURCE_IDENTIFIER));
     }
 
     @Test
@@ -295,8 +294,8 @@ public class CategoryReferenceResolverTest {
             .hasFailed()
             .hasFailedWithThrowableThat()
             .isExactlyInstanceOf(ReferenceResolutionException.class)
-            .hasMessage("Failed to resolve custom type reference on CategoryDraft"
-                + " with key:'key'. Reason: Reference 'id' field value is blank (null/empty).");
+            .hasMessage(format("Failed to resolve custom type reference on CategoryDraft with key:'key'. Reason: %s",
+                BLANK_ID_VALUE_ON_RESOURCE_IDENTIFIER));
     }
 
     @Test
@@ -311,8 +310,8 @@ public class CategoryReferenceResolverTest {
             .hasFailed()
             .hasFailedWithThrowableThat()
             .isExactlyInstanceOf(ReferenceResolutionException.class)
-            .hasMessage("Failed to resolve custom type reference on CategoryDraft"
-                + " with key:'key'. Reason: Reference 'id' field value is blank (null/empty).");
+            .hasMessage(format("Failed to resolve custom type reference on CategoryDraft with key:'key'. Reason: %s",
+                BLANK_ID_VALUE_ON_RESOURCE_IDENTIFIER));
     }
 
     @Test